Modify /system and reboot instead of bind-mount and framework restart
Restarting framework doesn't seem to work on a lot of devices, so we now modify
the /system partition with a wrapper-script for surfaceflinger. This wrapper checks
for /data/system/sbs/enable and /data/system/sbs/permanent and if either of them
exists it surfaceflinger is started with an LD_LIBRARY_PATH setup to use our modified
version of libsurfaceflinger.so.

/data/system/sbs/enable is so that we can reboot again to recover if things doesn't work.